host: support gate activations before VPE starts.
The basic idea is to put EPs, receive buffers, and the heap in one
mmap'ed area and let the kernel know the address. This happens via the
VPECTRL_INIT syscall at the beginning. In general, RecvGate stores the
offset of the buffer within the receive buffer area on the host. When
configuring the EP, the kernel adds the base address of the receive
buffer area.