host: use polling again.
There is a weird behavior with sendto/ppoll in some cases, leading to
both threads being blocked in system calls. I suspect that this is a
lost-wakeup bug or similar in Linux, but I'm not sure and I'm unable to
create a minimal example that reproduces the problem. The problem
happens frequently in the rust pingpong benchmark when performing it in
a loop, though. Since it's not really important and I ran out of ideas,
I removed the blocking calls and use polling again.