rustm3: make VPE an owned field in Activity.
On the one hand, this introduces the restriction that we can use each
VPE at most once via run/exec, but we want to go there anyway. On the
other hand, it allows us to store the activity in some data structure
(with a reference in the activity we can't keep the activity longer than
the scope of the VPE). A nice side-effect is that the new API allows
immutable and mutable access to the VPE during the lifetime of the
activity, but forbids another run/exec, which is exactly what we want.