rustunittests: use VFS::stat instead of file writes in VPE-stop test.
The problem is that the child shares the meta session with the parent
in this test and the meta session has a limited number of slots for
files. Therefore, killing the child leaves the file open and therefore,
all slots are occupied after a few loop iterations.

Previously, this worked because the kernel sent a close message to the
service upon the revoke. The kernel does no longer do that and the
resource manager doesn't know about these sessions.