rust: fixed pointer problem in EpMux.
We cannot simply store a pointer to the gate in EpMux to modify it
later when we multiplex EPs, because rust might move gates around,
making the pointer invalid.
Therefore, the new solution is to remember which EP is currently
owned by which selector (identifying the gate). Thus, whenever we
want to use a gate we check whether it had a EP once and whether
this EP is still owned by the gate's selector.