root: handle service requests asynchronously.
This is required, because some service requests (e.g., shutdown) might
lead to a request to root, which would deadlock if done synchronously.
Another reason is, that we do not have to trust the services that much,
because they cannot DOS attack the root task anymore by not replying to
these requests. And finally it helps to keep root more responsive.