pipes: introduced meta session to create pipes.
Previously, every new session at the pipe server created a new pipe and
thus involves the resource manager. There is not really a point in that,
because the memory for the pipe is provided by the client anyway and the
number of pipes can be defined when creating the (now existing) meta
session.