host: fixed race in VPE::{run,exec}.
The problem was that the child VPE needed some time to start the DTU and
thereby be ready for communication. If somebody else tried to
communicate with the child VPE before that point, it failed.

Thus, I'm using another wait-signal pair between parent and child to
signal the parent as soon as the child is ready for communication and
block the VPE::{run,exec} call until that point. For exec, the fd is
written to a file and the new application signals the completion.