Added subsys as a first second-level resource manager.
The idea is that the subsystem's resources are restricted as a whole,
but within the subsystem there are no restrictions. That means, which
VPE within that subsystem allocates how much memory, for example, does
not matter. This is primarily intended for tests that register own
services. In some cases, the number of services or sessions varies, so
that it is hard to restrict with our current config language. And in
these cases we actually do not care who opens how many sessions within
the subsystem.

For that reason, I moved the resource-manager part into a library and
use the library in root and subsys. It is probably not sufficiently
general for other subsystems, but we can extend it later.