Refactored WorkLoop.
Previously, the workloop was a global object that did some magic in the
background. For example, the RecvGate added a WorkItem to the WorkLoop
in the start method. For the programmer, the connection between those
was totally unclear. That is, it was not clear that calling
RecvGate::start requires using WorkLoop::run afterwards.

Now I changed WorkLoop to be created by the program and be passed to
the places where it's needed. This makes it a bit less convenient, but
also much more clear for the programmer.

Furthermore, I removed the sleep handler from the WorkLoop again.
Instead, run() now only puts the CU to sleep if there were no events.
Otherwise, it calls all WorkItems again to check whether there is
something to do. This makes sense in general and is sufficient for the
custom sleep handler that we had before.