kernel: ensure the VPE is stopped when destructing it.
This can occur if the VPE has not run anything and the owner doesn't
call stop explicitly. We have to make sure that it is stopped to flush
the PTEs to memory. Otherwise we might leak memory, because some PTEs
are not in memory yet and we don't find them in AddrSpace::remove_pts().