Added accounting of kernel memory.
Kernel memory is represented by a new type of kernel object, the
KMemObject. Each KMemObject stores a quota that guarantees access to a
certain amount of kernel memory.
We currently account kernel memory for kernel objects, capabilities,
page tables, and memory allocated by the kernel for mappings (e.g.,
rctmux's data segment). The rest cannot be attributed to a specific VPE
and therefore the kernel needs to statically calculate the required
memory for it (FIXED_KMEM). Everything else is passed to the root VPE as
a KMemObject. The kernel offers a system call to determine the remaining
quota of a KMemObject and to derive a new KMemObject from it. The latter
moves some of the quota to the new object and passes it back on revoke.
On VPE creation, the KMemObject to use for the VPE is specified.

For each VPE that root starts, the kernel memory can be specified
(kmem=...). If it is specified, root derives a new KMemObject with the
desired amount for the VPE. Otherwise, root equally distributes the
available kernel memory among the VPEs. Additionally, root can be
instructed to use the same KMemObject for all VPEs that have no
restriction (samekmem).