Use in-kernel paging on PE type C.
Previously, the kernel run with the 1:1 translation set up by gem5 when
using PE type C. Now the gem5 loader sets up x86 PTs for the kernel
accordingly. This requires the kernel to handle TLB misses for the DTU.
For that reason, the kernel supports interrupts and exceptions now.
Additionally, the kernel dynamically extends the kernel heap on
PE type C now.