Added a semaphore kernel object.
Previously, we used message passing channels to synchronize VPEs or
service-session dependencies. The former worked well if the VPEs knew
each other and the latter worked well for server-client relations.
However, we lacked a good synchronization mechanism for VPEs that don't
know each other.

Therefore, I introduced a semaphore kernel object. Anyone can create a
new semaphore and can perform up/down. Most importantly, root supports
the creation of semaphores and also the 'attachment' to semaphores.
With this, we can create a 'global' semaphore and attach VPEs that do
not know each other to it to synchronize them.