host: handle crashes of foreign childs properly.
Previously, we only detected crashes of direct childs of the kernel,
that is root at the moment. Now we have a dedicated communication
channel from each child to the kernel to notify the kernel about crashes
so that it can kill the VPE and notify the parent in the expected way.