Allocate EPs from PEMux, handed out as capabilities.
Previously, all VPEs got EP capabilities for the DTU EPs from the start.
Now only VPEs on accelerators (without PEMux) get them from the start,
all others need to allocate them. The allocation is done indirectly
over the resource manager.

The resource manager should be the only one that has the capability to
allocate EPs on shared PEs. It does so via a system call, upon which the
kernel sends an upcall to the PEMux on the target PE. PEMux finds a free
EP and tells the kernel the EP number via the reply. Upon receiving the
reply, the kernel creates a new EP capability for that EP and gives that
to the resource manager. The resource manager gives the cap to the
requesting application, which can then either perform the activate
system call (e.g., for a RecvGate) or pass it to someone else (e.g.,
m3fs).

A few things are still preliminary:
- the kernel expects a VPE capability for the EP allocation. In the
  future we have to use something else that prevents a VPE to just do
  that for itself, bypassing the resource manager.
- the resource manager has to actually enforce the limits on EPs on
  shared PEs.
- VPEs running without PEMux on a PE should get their EP caps from the
  start as before.