Added PE capabilities.
The idea is that root receives capabilities for all user PEs in the
system and can delegate them to its child upon request. To create a VPE,
applications need to pass a PE capability to the kernel.

Furthermore, a PE cap holds an endpoint quota and the alloc_ep system
call expects a PE cap to take the endpoint from. Initially, all PE caps
have all available EPs in their quota. The system call derive_pe can be
used to create a new PE capability with a subset of the EPs transferred
from the old one to the new one.

Currently, the resource manager has no real policy on PEs and their EPs,
but just hands out one PE at a time with all available EPs.

Note also that the {alloc,free}_ep() calls to the resource manager have
been removed, because the alloc_ep syscall now expect a PE capability.