Renamed "fft" accelerator to "copy" accelerator.
That's because it never was a real FFT accelerator, but merely took (by
default) the time an FFT would have taken based on ALADDINs estimation.
In reality it simply copied input to output and I customized the time it
accounted for that anyway. For that reason, we simply call it copy
accelerator now.