Removed endpoint multiplexing and indirect access to EPs.
Instead, we now assume that we always have enough EPs based on an
upcoming change that stores EPs in memory. Additionally, applications
on shared PEs can now again directly access their endpoints, because of
another upcoming change that will let the DTU know about the owner of
an EP.

Since we will have a large and variable number of EPs, the kernel does
no longer provide the application with EP caps for all its EPs upfront,
but all EPs have to be allocated via a system call.

Now that we no longer have reply header, but use EPs for replies, the
kernel needs to manage reply EPs just as it manages the other EPs. For
hardware simplicity, reply EPs have to be contiguous. To avoid the
complications of a generic EP-Range cap, I decided to keep simple EP
caps and optionally have an a set of EPs attached for replies.

Note also that the pager needs quite some EPs in some scenarios. At the
moment, the pager needs a bit more than necessary, because it doesn't
remember to which regions of a file it already has access. This could be
improved. However, in general, the pager needs (APP_MEM / BLOCK_SIZE)
memory EPs in the worst case.