Use a DTU-based communication channel between PEMux and kernel.
Previously, we had a shared-memory and polling-based communication
channel to notify PEMux about starts and stops of VPEs. I've replaced
that now with a DTU-based communication channel that is handled
asynchronously by the kernel. However, the old mechanism is still used
to invalidate TLB entries, because this is more difficult to do
asynchronously.