kernel: added workaround for pager EP invalidation.
We currently have fixed EPs to the pager for simplicity. For that
reason, they are not part of the "free user EPs" and therefore are not
invalidated by default. I've now introduced a special handling for those
EPs by creating EPObject instances for them and linking the gates to
them. The EPObject instances are remembered in the VPE for a later
destruction.