gem5: extended EPs to 16 bit and reduced addresses to 32 bit.
We will need more endpoints on a single PE in the future because of PE
multiplexing. However, we will not use 64-bit cores in the foreseeable
future. For that reason, I decided to only use 32-bit addresses, even
though we use x86_64 cores in simulation.

Note also that we are still using 64 bit addresses for memory EPs
(address and size). Additionally, I now use the offset register for the
offset of memory transfers. This is primarily due to PTE creation at the
moment, because these are located at a rather high address in the
virtual address space.