Merged init and pager.
In other words, the pager is now always the parent of the VPE it pages.
This has the advantage that it simplifies memory accounting, because we
can specify the memory quota for the pager and its entire subsystem.

Previously, it was more difficult because we would have to specify a
quota as a session parameter for each client and also the sum of it and
something else for the pager itself.

Additionally, init was more or less useless, because its only purpose
was to create a child with a pager and mount the root file system. Thus,
merging that with the pager makes also sense, because we need one
component less in the system.