Added memory pools/quotas and requests of physical memory ranges.
First, we now assign a separate memory pool to each child root creates.
These pools are, by default, equally large. The config option
"mem=<size>" can be used to request a smaller or larger pool.

Second, applications now need to request access to specific physical
memory ranges (e.g., FS image) via the config option
"memat=<addr>:<size>". In that case, this memory is added to their pool
and can be requested later. This can only be done for reserved memory,
that is, the memory that is not used for other allocations.