kernel: create dummy devices for IDE/NIC on host.
Otherwise, root complains that they are required by the config file,
but don't exist. They are not actually used in the end, but the kernel
needs at least to tell root that they exist.

The alternative would be to make them optional or have different configs
for host and gem5, which I think is worse.