Describe configuration via XML instead of magic cmdline arguments.
XML is more user friendly than the custom syntax we had before, is still
relatively easy to parse, and is particularly nice to describe PE
sharing.

For that reason, we now pass the XML config as a separate boot module to
root. Additionally, we use xmllint (if present) to validate the XML file
against an XSD (misc/boot.xsd) before starting the system to catch
simple mistakes early.