kernel: cleaned up capability revoke.
Previously, it was a bit unclear why and when revoke() was used to do
cleanup and when the destructor of the kernel object was used for that
purpose. Additionally, there was revoke() and late_revoke().

I've now reduced the number of times we need to do cleanup in revoke
() and added comments that explain the reason for that. Additionally,
late_revoke() was removed and revoke() is now always called after all
child/sibling capability have been removed.