pemux: put fixed entry into TLB to disable IRQs when sending messages.
The idea is that we put the "messages page" into the TLB at the
beginning to be able to disable IRQs when sending messages later. To
achieve that, the kernel sends an initial dummy upcall to PEMux and
PEMux replies to that message. This reply leads to a TLB miss and thus a
core request, upon which we put the fixed entry into the TLB.
Afterwards, no further TLB misses occur for sending messages, allowing
us to leave IRQs off most of the time, which greatly simplifies PEMux,
because we do not have to worry about nested IRQs so much.