pemux: use privileged command to create fixed TLB entry.
PEMux needs to insert a fixed entry into the TLB for every address
space. Thus, doing that once at the beginning does not work. For that
reason, we cannot easily do that by causing a TLB miss and using our
translate IRQ handler to put that entry into the TLB.

An alternative would be to have global TLB entries, valid for all
address spaces. But this sounds also expensive and having a dedicated
command to insert an entry seems like the simpler and cleaner solution.
However, it requires another privileged register to hold all the
arguments we need for the TLB insert command.