pemux: don't wait in PEMux, but schedule another VPE.
That means, in the sleep PEXCall, we do no longer use the TCU to wait
for a message, but block the VPE and schedule a different one
(idle atm). As soon as a message arrives, we switch to that VPE and, in
this case, just exit to userspace.

For pagefaults, we just send the message and then block the VPE.
Additionally, we register a function to call when we're scheduled next.
If a message arrived for us, the registered function checks whether we
have a reply from the pager. If so, we finish the pagefault handling
and return to userspace. Otherwise, we schedule again.

In consequence, we have no nested interrupts anymore, which greatly
simplifies the code and the mental model.