gem5: don't reset PE for new VPEs, but let PEMux flush the cache.
Since we want to have multiple VPEs on one PE in the near future, we
can no longer reset the PE on new VPEs. However, we still need to and
want to flush+invalidate the cache on VPE terminations to make sure that
new VPEs don't see the old state and to make sure that new VPEs have a
fresh view on memory.