Added auto-close feature to session.
When creating a session, the server can ask the kernel to send the close
message to the server as soon as all derived session caps have been
revoked. This is important for services such as the pipe service, where
the other side of the pipe should receive EOF if one side crashes, for
example.