kernel: revoke caps on exit instead of on destruction.
To make auto-close sessions work properly the kernel revokes now almost
all capabilities on exit instead of on destruction. That is required if,
for example, the parent communicates with its child, because in that
case the parent is the one that waits for the child when accessing the
pipe *and* is the one that needs to destroy the child. Thus, revoking
caps on destruction is not sufficient. However, after the child exited,
the caps are mostly useless anyway. In theory, the parent (e.g.) could
extract caps after the child exited, but was never used so far.

Note that the kernel does not revoke all caps on exit, because we still
want to keep the VPE cap, memory cap, kmem cap and PE cap around.