Added disktest program and test file system.
The problem is that if we want to run that in our test system, we cannot
generate a disk every time, because that requires root privileges. So,
we don't want to use the normal file systems, because they contain all
binares, which need to be regenerated after every change to M3. Instead,
we pre-generate a file system with specific test data and load all apps
in the test via boot module.