pemux: run idle VPE in user mode and don't put it on ready list.
Running it in user mode removes some special cases for idle. The
downside is that we have to make the idle loop user-accessible, which
implies that we have to know exactly what code it touches. The easiest
way is to implement it in assembly, giving us full control.
Unfortunately, we have to change the code as soon as the TCU ABI
changes.