Let PEMux map the environment, not the pager.
This way, we get rid of the separate PE info page (for the field
"shared") and can remove the special case for the pager when loading
VPEs. The problem before was that PEMux couldn't write to the
environment to set the field "shared", because the environment was
mapped by the user and was subject to copy-on-write. Thus, after a VPE
clone, PEMux pagefaulted when trying to set "shared". For that reason I
added another page, mapped by PEMux, only for "shared".

The much better solution is to move "shared" back into the environment
and simply let PEMux map the environment. We don't need copy-on-write
for it and it always needs to be mapped in a specific way. Therefore,
PEMux can always safely write to it. Additionally, when initializing the
environment we can always access it, even when we are the pager. Thus,
the special case to do that in the pager could be removed with this
change.