rustkernel: use RefCell less coarse grained.
So far, we put every kernel object into a RefCell, which made it really
hard to make sure that we do not violate the ownership model at runtime.
Additionally, it had some impact on performance.

Thus, this has been changed now to only use RefCell/Cell within the
kernel objects for the properties that actually need to change.