pemux: use hlt/wfi in idle loop.
Since we only run this loop with CUR_VPE=idle and idle does not receive
messages, we always get an interrupt if there is a message. So, using
hlt/wfi is perfectly fine. Additionally, it turned out that wfi can be
used in user mode on RISC-V if status.tw is 0. On x86, hlt is actually
privileged, but gem5 doesn't care - good for us.

Actually, this change is required, because we now have timer interrupts
and the TCU-based sleep puts the core to sleep even if there is a
pending timer interrupt. That is, if things go bad we wakeup because of
the timer interrupt, but still execute a couple of instructions in the
idle loop until the interrupt is finally injected. In this case, we
might execute the SLEEP command again, putting the core to sleep;
potentially for ever.