arm: convince the compiler to use our memset for zeroing memory.
The compiler tries to be smart and replaces calls to memset with calls
to __aeabi_memclr. I guess, it uses __aeabi_memclr instead of, for
example, __aeabi_memclr8, because memset cannot assume anything about
the alignment. Even if the address and size are page-aligned constants,
it ignores that fact and still calls __aeabi_memclr. Unfortunately,
__aeabi_memclr zeros the memory byte wise m(.

So, the workaround is to implement this function ourself and calling a
function the compiler doesn't know so that it doesn't replace that with
a call to __aeabi_memclr.