Reworked receive buffers.
With SPM, nothing really changed. But with caches+VM, we now have a
large virtual area for receive buffers and fill it on demand. The first
page is used for the standard receive buffers. Whenever a new RecvGate
is created, we allocate memory from our resource manager for it and map
it to a free space in that virtual area.

Besides the increased flexibility, it is a preparation for making
receive EPs use a physical address instead of a virtual address,
requiring a physically contiguous receive buffer. Now that PEMux only
allocates one frame for the standard receive buffers and for all others
we allocate physically contiguous memory individually, it becomes much
easier.