gem5: decouple core requests from transfers.
That means, we have an independent queue of pending core requests, which
are issued one at a time by writing to CORE_REQ and completed as soon as
the core writes to CORE_REQ again.

Additionally, the foreignRecv core request does no longer pause the
transfer, but is just sent as a core request to software after the
transfer as a notification. This prevents issues if, for example, PEMux
sends a pagefault message to the pager with interrupts disabled.