libm3: poll for some time in TCUIf::receive.
This is not strictly necessary, but in some situations, a VPE is
interrupted by a message for another VPE and is then put into the ready
queue. Sometimes this was "unnecessary", because the old VPE would have
went to sleep anyway, but because it was interrupted, we have a ready
VPE. Thus, PEMux reports that the new VPE should inform PEMux in case it
wants to wait for a message and this does not change until we switch to
the old VPE again (e.g., if the budget of the new VPE is depleted).

While this behavior is correct, it has the problem (besides the
performance issue itself), that performance in our PE-sharing benchmarks
becomes more or less unpredictable. For that reason, I introduced
polling to delay yielding a bit, which should cover most cases in our
benchmarks.