Fixed problems with stream processing accelerators.
Most importantly, the stream processing accelerators need to distinguish
between the end of one run and a new run. At the end of a run, it needs
to ignore incoming messages. However, at the beginning of a new run, we
CANNOT ignore incoming messages. Thus, we need a kind of signal in
between.

The fix for now is to reset accelerators  when the VPE was destroyed,
which clearly marks the end of one run. And on the start signal, we
don't reset the state anymore, but simply set the start signal.