Make remote memory accesses always refer to physical memory.
The reasoning is given in the gem5 commit. This has a few consequences
for M3. At first, VPEs do no longer receive a memory capability for
their entire address space. Instead, there is a new system call that can
be used to get access to a physically contiguous portion of the address
space. Whether it is contiguous is defined by having been mapped by a
single create_map call.

There is one special case: the environment is mapped by PEMux using a
physical address of its choosing. Since we need to initialize that from
the loader, PEMux provides a new upcall to translate an address. The
kernel uses that to insert an entry into the map capabilities, so that
it can be accessed by applications. The translate call does only work
for a single page, which is also the reason I reduced the size of the
environment to a single page for now. Note that this translate upcall
will also allow PEMux to put receive buffers at an arbitrary physical
location again.