kernel: ask PEMux for standard receive buffer address.
Since we now have the translate upcall anyway, the kernel can also use
it to get the physical address of the standard receive buffer. In that
way, PEMux doesn't need to dedicate a range of frames for that purpose
(so that the kernel can calculate the address itself), but can use any
physical address.