Reworked service/session creation.
Preparing for the upcoming resource manager changes, I've reworked the
way services and sessions are created. The idea is to bind a session
quota to the SendGate used to communicate with the server, maintained by
the server. This way, a resource manager can create a new SendGate to a
server with a quota for a specific number of sessions, so that each
resource manager can open/close sessions autonomously, but in a
secure fashion.

The server maintains a list of "session creators", each with their own
SendGate and a session quota. The ID of the session creator is used as
the label for the SendGate, so that the creator cannot forge it. The
server reduces/increases the session quota on every open/close for the
corresponding creator. Additionally, a new system call and service call
for deriving a new session creator has been introduced. The derive_srv
system call takes an existing service cap and creates a new one with a
new session creator id and a new SendGate with a specified number of
sessions. Within derive_srv the kernel uses the derive_crt service call
to ask the server whether sufficient quota is available and which
creator id and SendGate to use. The service cap and SendGate can be
passed to someone else, giving them access to at most N sessions.