Added get_sess system call.
This will be used shortly by the resource manager to get a session
capability for one of the sessions that it created. Previously, this
wasn't necessary, because the resource manager was always the direct
parent of the service and thus had complete access to the VPE 
(allowing us to just take the capability).

Now that we can have multiple session creators, the resource manager
that wants to open a new session might not be the direct parent of the
service. Therefore, I introduced a system call that allows it to get
the session capability after creation, so that it can be passed to the
client (actually, in the same call, because it takes a VPE argument).

Note also that there is no other way for the resource to get to that
capability, because it does not necessarily have a session at the
service, so that it can not exchange capabilities with the service.