Redesigned resource manager(s) and pager.
The idea is to allow nesting of subsystems and their description via one
config file. For that reason, both root and pager are now a resource
manager in the sense that they provide their children access to
resources upon request.

To allow arbitrary nesting without performance penalty, a subsystem
(that is, an app with children) receives the resources for its children
at startup. For that we use the boot info, which is already passed from
the bootloader to the kernel and from the kernel to root. Thus, every
layer in this nested system receives its resources from the parent
layer. The resources are currently boot modules, PEs, memory, and
services.

Supporting the "pass down" of services was particularly challenging,
because so far a single resource manager was responsible for a service
and the service trusted this resource manager to quite some degree. For
that reason, I reworked the service/session creation (see separate
commit), so that multiple resource managers can create sessions at the
same service without interfering with each other.

The config file is passed to root as a boot module (as before), which
parses it and checks its validity (availability of required resources,
for example). For a subsystem, it extracts the part of the configuration
file that describes it and passes it as a boot module to the subsystem.

With that in place, we can now also isolate applications within a
subsystem and use all other benefits of the configuration file.