resmng: allow specification of session quota for subsystems.
Previously, we collected dependencies for subsystems by walking through
the children and collecting the services that are used. We then
calculated the session quota for the subsystem simply based on the
number of childs that use a specific service. This results of course
easily into a waste of session space.

Therefore, the idea now is that the session creator quota can be
specified for subsystems to use a fixed number of sessions. If it is not
specified, the sessions are split equally among the subsystem and the
other childs at the same level.

This is not perfect, because when equally distributing sessions we still
make an assumption about the number of "subsessions" created by each of
these sessions. Their number can be specified in a service-specific way,
like with "files=n" as an argument to m3fs, but this is independent of
the session quotas in the config. However, that's maybe something for
later.