Added support for the FPGA.
We now have a new target called "hw". Since we aim for binary
compatibility, we group gem5 and hw together by introducing a
"platform". The platform is either host or kachel. This is actually not
user facing, but just the way the implementation handles targets in the
same way, while still allowing to distinguish between them.