thread: fixed bug in stop.
If there are sleeping threads, we need to switch to those. This is
because we finally need to return to the main thread, which exits
instead of calling stop. Thus, if in the rustkernel the last VPE exited
with a worker thread, we need to switch through the threads until we hit
the main thread to finally exit.