rustkernel: fixed memory-safety issue in revocation.
We cannot remove a capability from one Treap and insert it in another.
The problem is that they have pointers to each other and removing a cap
from the Treap free's the node they are stored in. That is, the pointers
are invalid afterwards.

As it turned out, there is a much easier way to solve the problem. We
can revoke all capabilities on exit if we don't inherit the VPE cap of
the parent from the child's VPE cap. This was always weird for other
reasons as well. Thus, we are now just giving both a "free standing"
VPE capability and make sure to not remove the VPE from the VPEManager
if the own VPE cap is revoked.