librustbase: prevent memcpy calls for Treap::insert.
The first problem was that I allocated the node without boxing it first,
and boxed it afterwards (doh!). The second problem was the compiler
created the value outside the function and then copied the value into
the node inside the function.

By using a separate and small function to do the allocation of the node
from key and value and calling the larger insert function, the can
convince the compiler to create the value in place and thereby prevent
all memcpy calls.