Use opcode for all pager requests, including the obtain/delegate based.
Previously, we distinguished between different obtain/delegate requests
on a case-by-case basis by looking at the argument or capability count
or similar. This was fine initially when we just had one or two of
those requests, but now we need a more systematic way for it to improve
readability, maintainability, and debugability of the system.