resmng: support mount specification.
Previously, the pager always passed the root filesystem to its children,
which is of course not the way to go. I've now introduced a new tag called
"mount" for apps in the config, which tells the resource manager to
mount a specific filesystem at a specific path. Since the resmng
potentially needs to open a new session, the filesystem name refers to
a session that the resmng can open.

Note that this is only done by the pager, because root doesn't have
filesystem access.