Switched from scons to our own meta build system.
While ninja as a build system backend is great, the scons-to-ninja
translator that we had so far was a bit of a pain. At first, the
description is scattered among different places (scons + translator) and
the translator needs to parse the dependencies generated by scons, which
makes it sometimes hard to get things right. Second, for every new
command we want to add to the build, the translator needs to be touched.
Third, scons is still run once in a while, which costs a significant
amount of time.

Therefore, I've now decided to replace scons with our own meta build
system. It's still written in Python and you can think of it as a very
very simple version of scons, that does not build anything, but merely
generates a ninja build file. This is because some concept are inspired
by scons (e.g., the Env). Since ninja build files are very simple,
building a generator is quite simple, too.

What we have now is the src/tools/ninjagen.py, containing generic
abstractions to generate a ninja build file. These could be used in
other projects as well, although I didn't test that. On top of that, we
have configure.py and a build.py in every directory that contributes to
the build. The configure.py and the build.py files are M³-specific.

For the user, nothing has really changed. The shell script b is still
used for building, inspecting, debugging, running, and so on, which
automatically (re-)configures (generating the ninja build file) and
builds M³ (by calling ninja).

In summary, having our own meta build system turned out to be better in
pretty much all dimensions. It's about two orders of magnitudes faster
than scons during configure, it's 500 lines less code in total, it's
much easier to use because we got rid of one indirection, and we can do
whatever we like to fit our needs. The only downside is that we don't
have all features of scons anymore, which were sometimes handy
(e.g., checking for header file existence), but I think we can live with
that.