rust: always disable FPU instructions on x86_64 for now.
The problem is that if we want to build all rust crates with a single
workspace, all crates need to be built in the same way. In theory, we
could work around this limitation by removing PEMux from the workspace,
and using a separate target description for PEMux, as we did so far. But
this causes other trouble and I think for x86_64 it's just not worth it.