Use upcall for the completion notification of the derive_srv syscall.
As it turned out, handling derive_srv synchronously can lead to a
deadlock in some cases. This is because the server that receives the
derive_crt service call on behalf of this syscall might have sent a
request to the resource manager before. So, the resmng would wait
(indirectly via the kernel) for the server and the server would wait for
the resmng.

There does not seem to be a better solution than making the derive_srv
call asynchronous to prevent this deadlock. Thus, the kernel now replies
immediately to acknowledge that the request has been sent to the server,
and notifies the caller later via upcall about the completion of this
request.