kernel: invalidate reply EPs on SendGate deactivation.
Whenever we deactivate a SendGate, it could have been used before to
send a message to the receiver without receiving its reply yet.
Typically, this happens if the sender crashes or is force-destructed.
Since a later reply could cause trouble (e.g., might be received by a
different VPE), the kernel has to ensure that no further reply is
possible. Thus, it needs to invalidate all the reply EPs that give
credits back to the send EP we are about to deconfigure.