kernel: use naming convention to make thread switches visible.
Since the performance implications of async+await turned out to be too
severe, I decided to go for a compromise: name functions *_async if
they perform a thread switch or call *_async functions and use a tool
to notice deviations from this convention.

With this, the name of a function makes it apparent where thread
switches might happen, so that we can easier make sure to be prepared
for them.

The tool (src/tools/async-check.py) is far from perfect, because it uses
a simple regex-based approach to find function definitions and calls,
but I consider it sufficient for our purposes. For example, it ignores
strings and comments, potentially leading to false positives or parsing
failures if a string/comment contains a function signature or call.