m3fsrs: improved safety of metabuffer.
At first, I introduced MetaBufferBlockRef which stores a link to a
MetaBufferBlock during its existence and thus replaces the manual
counting of block references in Request. For that reason, Request was
removed.

Second, we do no longer hand out references with a static lifetime to
metabuffer blocks, which enables more compile time checks.

Third, we do no longer allow mutable access to DirEntry's while
iterating over them. In other words, the DirEntryIterator only hands
out immutable references to DirEntry. This prevents subtle mistakes
while using the iterator, as could have happened before. The downside
is that we need to implement the iterator manually for the link
operations in order to be able to modify DirEntry.