m3fsrs: improved INode handling.
Unfortunately, we cannot distinguish between immutable and mutable INode
references, because (for example) at some points we need to pass the
same mutable reference twice to a function. Additionally,
INodes::get_extent needs to always take a mutable INode reference just
in case the "create" argument is true and we need to create a new
extent.

For that reason, we just have INodeRef which provides both immutable and
mutable access to an INode. Immutable access is performed via Deref and
mutable access needs to be explicitly requested by calling as_mut.

Additionally, instead of storing a RefCell, we store a raw pointer which
simplifies a few things. We know that this pointer is valid during the
lifetime of INodeRef, because INodeRef keeps a reference to the
underlying metabuffer block.