rust: reworked serialization framework.
Apparently, with the previous approach based on traits for Sink/Source
and multiple implementations, the compiler was unable to see that all
serialize/unserialize calls can actually be inlined, because all types
are known at compile time and therefore, no dynamic dispatch is
needed.

For that reason, I've rewritten that now to use a single Sink and a
single Source struct for which we implement serialize/deserialize
support for various data types. That worked well for both normal
messages and arguments to delegate/obtain, because both can be
implemented based on slices. For the VPE state serialization we use a
different implementation called StateSerializer based on a vector.
Fortunately, we don't need support for all kinds of data types here,
but integers and strings are sufficient.